requirements.txt

flask
openai
wikipedia
python-dotenv
streamlit
aiohttp